+++ vitest.config.ts
@@ -5,6 +5,7 @@ export default defineConfig({
   plugins: [react()],
   test: {
     environment: "jsdom",
+    exclude: ["**/node_modules/**", "**/dist/**", "**/src/tests/e2e/**"],
     globals: true,
     setupFiles: ["./src/tests/setup.ts"],
   },